+++ src/main/kotlin/io/github/krozov/detekt/koin/KoinRuleSetProvider.kt
@@ -3,11 +3,17 @@ package io.github.krozov.detekt.koin
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
+import io.github.krozov.detekt.koin.moduledsl.ConstructorDslAmbiguousParameters
 import io.github.krozov.detekt.koin.moduledsl.DeprecatedKoinApi
+import io.github.krozov.detekt.koin.moduledsl.DuplicateBindingWithoutQualifier
 import io.github.krozov.detekt.koin.moduledsl.EmptyModule
+import io.github.krozov.detekt.koin.moduledsl.EnumQualifierCollision
+import io.github.krozov.detekt.koin.moduledsl.GenericDefinitionWithoutQualifier
 import io.github.krozov.detekt.koin.moduledsl.MissingScopedDependencyQualifier
 import io.github.krozov.detekt.koin.moduledsl.ModuleIncludesOrganization
+import io.github.krozov.detekt.koin.moduledsl.ParameterTypeMatchesReturnType
 import io.github.krozov.detekt.koin.moduledsl.SingleForNonSharedDependency
+import io.github.krozov.detekt.koin.moduledsl.UnassignedQualifierInWithOptions
 import io.github.krozov.detekt.koin.scope.CloseableWithoutOnClose
 import io.github.krozov.detekt.koin.scope.FactoryInScopeBlock
 import io.github.krozov.detekt.koin.scope.KtorRequestScopeMisuse
@@ -61,6 +67,12 @@ public class KoinRuleSetProvider : RuleSetProvider {
                 MissingScopedDependencyQualifier(config),
                 DeprecatedKoinApi(config),
                 ModuleIncludesOrganization(config),
+                UnassignedQualifierInWithOptions(config),
+                DuplicateBindingWithoutQualifier(config),
+                GenericDefinitionWithoutQualifier(config),
+                EnumQualifierCollision(config),
+                ConstructorDslAmbiguousParameters(config),
+                ParameterTypeMatchesReturnType(config),
                 // Scope Management Rules
                 MissingScopeClose(config),
                 ScopedDependencyOutsideScopeBlock(config),