+++ src/test/kotlin/io/github/krozov/detekt/koin/KoinRuleSetProviderTest.kt
@@ -18,6 +18,6 @@ class KoinRuleSetProviderTest {
         val ruleSet = provider.instance(config)
 
         assertThat(ruleSet.rules).isNotEmpty()
-        assertThat(ruleSet.rules).hasSize(32)
+        assertThat(ruleSet.rules).hasSize(38)
     }
 }